Specify uint256

+++ contracts/MasterCounter.sol
@@ -76,7 +76,7 @@ contract MasterCounter is ICounterDeployment, ReentrancyGuard {
             gasForDestinationLzReceive
         );
 
-        (uint messageFee,) = endpoint.estimateFees(
+        (uint256 messageFee,) = endpoint.estimateFees(
             _dstChainId,
             address(this),
             _payload,
